Fractional scaling

+++ src/java.desktop/unix/classes/sun/awt/wl/WLGraphicsEnvironment.java
@@ -253,7 +253,7 @@ public class WLGraphicsEnvironment extends SunGraphicsEnvironment {
         }
     }
 
-    static double effectiveScaleFrom(int displayScale) {
+    static double effectiveScaleFrom(double displayScale) {
         return debugScaleEnabled ? SunGraphicsEnvironment.getDebugScale() : displayScale;
     }
